fixed splash screen file structure
