+++ src/components/SplashScreen/SplashScreen.css
